moved polygon manipulation to initialization, added plotting offsetted polygon
Moved offset and convex hull before bounding box creation of prepared polygon.
Offset polygon is now used for plotting.
Some skipped tiles seem to be present when offsetting, different buffer style may fix that.

Not happy yet with the coverage of the exterior line of the polygon, may be because of the bounding box is under/over-influenced by fov-size.