feat: added GridFromPolygon (#224)
* added GridFromPolygon

* style(pre-commit.ci): auto fixes [...]

* shapely for all logic, and refactored pydantic bit

* style(pre-commit.ci): auto fixes [...]

* feat: brought tile formation in line with other gridplans, and updated plotting

* style(pre-commit.ci): auto fixes [...]

* refactored docstrings

* style(pre-commit.ci): auto fixes [...]

* moved polygon manipulation to initialization, added plotting offsetted polygon

Moved offset and convex hull before bounding box creation of prepared polygon.
Offset polygon is now used for plotting.
Some skipped tiles seem to be present when offsetting, different buffer style may fix that.

Not happy yet with the coverage of the exterior line of the polygon, may be because of the bounding box is under/over-influenced by fov-size.

* style(pre-commit.ci): auto fixes [...]

* Added docstrings, removed some non functional pydantic annotations, enlarged initial bounding box

enlarged the initial bounding box slightly by fov, this ensures to-be-checked tiles will also be created on the extreme edges of the polygon.

* Import PrivateAttr to hide some attributes in the signature gridfrompolygon

* Fix private attributes

* style(pre-commit.ci): auto fixes [...]

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Talley Lambert <talley.lambert@gmail.com>